# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Document Analyzer & Court Simulator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Case {
  id: number;
  title: string;
  /** @nullable */
  defendantName?: string | null;
  /** @nullable */
  caseNumber?: string | null;
  /** @nullable */
  jurisdiction?: string | null;
  /** @nullable */
  notes?: string | null;
  hasAnalysis: boolean;
  hasMotion: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCaseBody {
  title: string;
  /** @nullable */
  defendantName?: string | null;
  /** @nullable */
  caseNumber?: string | null;
  /** @nullable */
  jurisdiction?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateCaseBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  defendantName?: string | null;
  /** @nullable */
  caseNumber?: string | null;
  /** @nullable */
  jurisdiction?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface CaseOrNull {
  case: Case | null;
}

export type DocumentDocumentType =
  (typeof DocumentDocumentType)[keyof typeof DocumentDocumentType];

export const DocumentDocumentType = {
  transcript: "transcript",
  police_report: "police_report",
  appeal: "appeal",
  motion: "motion",
  order: "order",
  affidavit: "affidavit",
  exhibit: "exhibit",
  policy: "policy",
  executive_order: "executive_order",
  other: "other",
} as const;

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  pending: "pending",
  analyzing: "analyzing",
  analyzed: "analyzed",
  error: "error",
} as const;

export interface Document {
  id: number;
  caseId: number;
  title: string;
  documentType: DocumentDocumentType;
  content: string;
  status: DocumentStatus;
  findingCount: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateDocumentBodyDocumentType =
  (typeof CreateDocumentBodyDocumentType)[keyof typeof CreateDocumentBodyDocumentType];

export const CreateDocumentBodyDocumentType = {
  transcript: "transcript",
  police_report: "police_report",
  appeal: "appeal",
  motion: "motion",
  order: "order",
  affidavit: "affidavit",
  exhibit: "exhibit",
  policy: "policy",
  executive_order: "executive_order",
  other: "other",
} as const;

export interface CreateDocumentBody {
  title: string;
  documentType: CreateDocumentBodyDocumentType;
  content: string;
}

/**
 * @nullable
 */
export type FindingPrecedentType =
  | (typeof FindingPrecedentType)[keyof typeof FindingPrecedentType]
  | null;

export const FindingPrecedentType = {
  BINDING: "BINDING",
  PERSUASIVE: "PERSUASIVE",
} as const;

export interface CrossCaseMatch {
  sourceDocumentId: number;
  sourceDocumentTitle: string;
  matchedPassage: string;
  explanation: string;
  relevanceScore: number;
}

export interface Finding {
  id: number;
  documentId: number;
  caseId: number;
  issueTitle: string;
  transcriptExcerpt: string;
  legalAnalysis: string;
  /** @nullable */
  precedentName?: string | null;
  /** @nullable */
  precedentCitation?: string | null;
  /** @nullable */
  precedentType?: FindingPrecedentType;
  /** @nullable */
  courtRuling?: string | null;
  /** @nullable */
  materialSimilarity?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  userNotes?: string | null;
  crossCaseMatches: CrossCaseMatch[];
  createdAt: string;
}

export interface UpdateFindingBody {
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  userNotes?: string | null;
  /** @nullable */
  issueTitle?: string | null;
}

export type CategoryColor = (typeof CategoryColor)[keyof typeof CategoryColor];

export const CategoryColor = {
  blue: "blue",
  yellow: "yellow",
  red: "red",
  pink: "pink",
  orange: "orange",
} as const;

export interface Category {
  id: number;
  name: string;
  badgeLabel: string;
  /** @nullable */
  description?: string | null;
  color: CategoryColor;
  createdAt: string;
}

export type CreateCategoryBodyColor =
  (typeof CreateCategoryBodyColor)[keyof typeof CreateCategoryBodyColor];

export const CreateCategoryBodyColor = {
  blue: "blue",
  yellow: "yellow",
  red: "red",
  pink: "pink",
  orange: "orange",
} as const;

export interface CreateCategoryBody {
  name: string;
  badgeLabel: string;
  /** @nullable */
  description?: string | null;
  color: CreateCategoryBodyColor;
}

export interface UpdateCategoryBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  badgeLabel?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
}

export type CourtSessionSimulationMode =
  (typeof CourtSessionSimulationMode)[keyof typeof CourtSessionSimulationMode];

export const CourtSessionSimulationMode = {
  direct_appeal: "direct_appeal",
  bangert_motion: "bangert_motion",
  postconviction_974: "postconviction_974",
  federal_habeas: "federal_habeas",
} as const;

export type CourtSessionStatus =
  (typeof CourtSessionStatus)[keyof typeof CourtSessionStatus];

export const CourtSessionStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  error: "error",
} as const;

export interface CourtSession {
  id: number;
  caseId: number;
  simulationMode: CourtSessionSimulationMode;
  skepticMode: boolean;
  expandedRecord: boolean;
  /** @nullable */
  pleaQuestionnaireNotes?: string | null;
  status: CourtSessionStatus;
  /** @nullable */
  verdictRating?: string | null;
  /** @nullable */
  verdictSummary?: string | null;
  /** @nullable */
  defenseWon?: boolean | null;
  /** @nullable */
  totalRounds?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateCourtSessionBodySimulationMode =
  (typeof CreateCourtSessionBodySimulationMode)[keyof typeof CreateCourtSessionBodySimulationMode];

export const CreateCourtSessionBodySimulationMode = {
  direct_appeal: "direct_appeal",
  bangert_motion: "bangert_motion",
  postconviction_974: "postconviction_974",
  federal_habeas: "federal_habeas",
} as const;

export interface CreateCourtSessionBody {
  simulationMode: CreateCourtSessionBodySimulationMode;
  skepticMode?: boolean;
  expandedRecord?: boolean;
  /** @nullable */
  pleaQuestionnaireNotes?: string | null;
  documentIds: number[];
}

export type CourtRoundStateStrength =
  (typeof CourtRoundStateStrength)[keyof typeof CourtRoundStateStrength];

export const CourtRoundStateStrength = {
  MINIMAL: "MINIMAL",
  MODERATE: "MODERATE",
  STRONG: "STRONG",
  OVERWHELMING: "OVERWHELMING",
} as const;

export interface CourtRound {
  id: number;
  sessionId: number;
  roundNumber: number;
  stateStrength: CourtRoundStateStrength;
  defenseBurden: string;
  stateArgument: string;
  courtCommentary: string;
  defenseResponse: string;
  createdAt: string;
}

export interface CourtSessionWithRounds {
  session: CourtSession;
  rounds: CourtRound[];
}

export interface Motion {
  id: number;
  caseId: number;
  /** @nullable */
  sessionId?: number | null;
  title: string;
  content: string;
  createdAt: string;
}

export interface CaseExport {
  case: Case;
  documents: Document[];
  findings: Finding[];
  categories: Category[];
  courtSessions: CourtSessionWithRounds[];
  motions: Motion[];
  exportedAt: string;
}