# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Legal Document Analyzer & Court Simulator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all cases
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseBody"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Not found
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseBody"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all custom issue categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a custom issue category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── COURT SESSIONS ──────────────────────────────────────────────────────────
  /cases/{caseId}/court-sessions:
    get:
      operationId: listCourtSessions
      tags: [court]
      summary: List court simulation sessions for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourtSession"
    post:
      operationId: createCourtSession
      tags: [court]
      summary: Create a new court simulation session
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourtSessionBody"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtSession"
  /cases/{caseId}/court-sessions/{id}:
    get:
      operationId: getCourtSession
      tags: [court]
      summary: Get a court session with all rounds
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Court session with rounds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtSessionWithRounds"
        "404":
          description: Not found
  /cases/{caseId}/court-sessions/{id}/run:
    post:
      operationId: runCourtSimulation
      tags: [court]
      summary: Run the full simulation (SSE stream — 4+ rounds then verdict)
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream of simulation rounds and verdict

  # ─── MOTIONS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Case:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        defendantName:
          type: ["string", "null"]
        caseNumber:
          type: ["string", "null"]
        jurisdiction:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        hasAnalysis:
          type: boolean
        hasMotion:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - hasAnalysis
        - hasMotion
        - createdAt
        - updatedAt
    CreateCaseBody:
      type: object
      properties:
        title:
          type: string
        defendantName:
          type: ["string", "null"]
        caseNumber:
          type: ["string", "null"]
        jurisdiction:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - title
    UpdateCaseBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        defendantName:
          type: ["string", "null"]
        caseNumber:
          type: ["string", "null"]
        jurisdiction:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        badgeLabel:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
          enum: [blue, yellow, red, pink, orange]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - badgeLabel
        - color
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        badgeLabel:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: string
          enum: [blue, yellow, red, pink, orange]
      required:
        - name
        - badgeLabel
        - color
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        badgeLabel:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]

    # ─── COURT SESSIONS ────────────────────────────────────────────────────────
    CourtSession:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        simulationMode:
          type: string
          enum: [direct_appeal, bangert_motion, postconviction_974, federal_habeas]
        skepticMode:
          type: boolean
        expandedRecord:
          type: boolean
        pleaQuestionnaireNotes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, running, completed, error]
        verdictRating:
          type: ["string", "null"]
        verdictSummary:
          type: ["string", "null"]
        defenseWon:
          type: ["boolean", "null"]
        totalRounds:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - caseId
        - simulationMode
        - skepticMode
        - expandedRecord
        - status
        - createdAt
        - updatedAt
    CreateCourtSessionBody:
      type: object
      properties:
        simulationMode:
          type: string
          enum: [direct_appeal, bangert_motion, postconviction_974, federal_habeas]
        skepticMode:
          type: boolean
        expandedRecord:
          type: boolean
        pleaQuestionnaireNotes:
          type: ["string", "null"]
        documentIds:
          type: array
          items:
            type: integer
      required:
        - simulationMode
        - documentIds
    CourtRound:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        roundNumber:
          type: integer
        stateStrength:
          type: string
          enum: [MINIMAL, MODERATE, STRONG, OVERWHELMING]
        defenseBurden:
          type: string
        stateArgument:
          type: string
        courtCommentary:
          type: string
        defenseResponse:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - roundNumber
        - stateStrength
        - defenseBurden
        - stateArgument
        - courtCommentary
        - defenseResponse
        - createdAt
    CourtSessionWithRounds:
      type: object
      properties:
        session:
          $ref: "#/components/schemas/CourtSession"
        rounds:
          type: array
          items:
            $ref: "#/components/schemas/CourtRound"
      required:
        - session
        - rounds

    # ─── MOTIONS ───────────────────────────────────────────────────────────────
...