# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Document Analyzer & Court Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface Case {
  id: number;
  title: string;
  /** @nullable */
  defendantName?: string | null;
  /** @nullable */
  caseNumber?: string | null;
  /** @nullable */
  jurisdiction?: string | null;
  /** @nullable */
  notes?: string | null;
  hasAnalysis: boolean;
  hasMotion: boolean;
  createdAt: Date;
  updatedAt: Date;
}