# lib/api-zod/src/generated/types/courtSessionWithRounds.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Document Analyzer & Court Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { CourtRound } from "./courtRound";
import type { CourtSession } from "./courtSession";

export interface CourtSessionWithRounds {
  session: CourtSession;
  rounds: CourtRound[];
}