# lib/api-zod/src/generated/types/createCategoryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal Document Analyzer & Court Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCategoryBodyColor } from "./createCategoryBodyColor";

export interface CreateCategoryBody {
  name: string;
  badgeLabel: string;
  /** @nullable */
  description?: string | null;
  color: CreateCategoryBodyColor;
}